Move latest_approval tests from 'approvals' to 'users' app
